% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixarZipsB3.R
\name{baixarZipsB3}
\alias{baixarZipsB3}
\title{Download dos zips com as cotações diárias de um ano.}
\usage{
baixarZipsB3(ano, out = "./Downloads", integrity = TRUE)
}
\arguments{
\item{ano}{Um inteiro representando o ano}

\item{out}{O caminho onde serão baixados os arquivos}

\item{integrity}{Checa a integridade dos arquivos .zip}
}
\description{
Faz o download dos arquivos zip, sem processá-los. Para extrair e carregar o conteúdo
para o R, use a função lerCotacoesB3. O FTP da B3 inclui cotações a partir de 1986.
Este pacote ainda não inclui um teste de integridade. É provável que arquivos com menos de 5mb não
foram baixados corretamente.
}
\examples{
baixarZipsB3(2017)
for (ano in 1990:1999) { baixarZipsB3(ano) }
}
