% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendario.R
\name{calendarioBR}
\alias{calendarioBR}
\title{Dias úteis Brasil}
\usage{
calendarioBR(dataInicio, dataFim)
}
\arguments{
\item{dataInicio}{data no formato "AAAA-MM-DD"}

\item{dataFim}{data no formato "AAAA-MM-DD"}
}
\value{
vetor de datas no formato "AAAA-MM-DD"
}
\description{
Constroi um calendário de dias úteis.São necessárias a data inicial
(AAAA-MM-DD) e a data final (AAAA-MM-DD)
Retorna um vetor com as datas.
}
\examples{
calendarioBR("2019-01-01", "2019-12,31")
calendarioBR("2020-06-10", "2020-06.25")
}
